requirements.txt
pandas==2.3.3
networkx==3.6.1
streamlit==1.55.0
plotly==6.6.0
pyvis==0.3.2
PyYAML==6.0.3
